dashboard/requirements.txt
pandas
altair
pylint
streamlit
python-dotenv
sqlalchemy
pymssql
boto3
mypy_boto3_s3